requirements.txt
asgiref==3.8.1
Django==5.1.7
gunicorn==23.0.0
packaging==24.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sqlparse==0.5.3
tzdata==2025.2
selenium==4.15.2
webdriver-manager==4.0.1
coverage==7.3.2
tblib==1.2.0
django-allauth==0.57.0
